requirements.txt
flask==2.3.2
pyjwt==2.7.0
cryptography==41.0.1